backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
requests
numpy
jinja2
markdown
python-multipart
